issue tickets 1 and two moved over my house data and voldy data to new folders and imported in to main js
