+++ public/main.js
@@ -1,67 +1,33 @@
-// import '../styles/main.scss'; // You have to import your styles for them to work. Comment in this line
+import '../styles/main.scss'; // You have to import your styles for them to work. Comment in this line
+import houses from '../utils/sample_data/houseData'; // importing house data
+import voldysArmy from '../voldyArray'; // importing voldy's Army
 
-const startApp = () => {
-  htmlStructure(); // always load first
-  header();
-  startSortingBtn();
-  events(); // always load last
+const createId = (array) => {
+  if (array.length) {
+    const idArray = array.map((el) => el.id);
+    return Math.max(...idArray) + 1;
+  }
+  return 0;
 };
 
 const students = [];
-const voldysArmy; // starts as an empty array
 
-
-const events = () => {
-  // get form on the DOM on button click
-  document.querySelector('#start-sorting').addEventListener('click', () => {
-    // put html elements on the DOM on click
-    form(); // form
-    filterBtnRow(); // filter buttons
-    studentAreas(); // students and voldy's army divs
-  });
-
-  // target expel buttons to move to voldys army
-  document
-    .querySelector('#student-container')
-    .addEventListener('click', (e) => {
-      if (e.target.id.includes('expel')) {
-        const [, id] = e.target.id.split('--');
-        const index = students.findIndex((student) => student.id === Number(id));
-
-        // move from one array to another
-        voldysArmy.push(...students.splice(index, 1));
-        // get both sets of students on the DOM
-        studentsOnDom('#students', students);
-        studentsOnDom('#voldy', voldysArmy);
-      }
-    });
-
-  // target filter buttons on Dom
-  document.querySelector('#filter-container').addEventListener('click', (e) => {
-    if (e.target.id.includes('filter')) {
-      const [, house] = e.target.id.split('--');
-
-        if (house === 'all') {
-          studentsOnDom('#students', students);
-        } else if (house) {
-          const filter = students.filter((student) => student.house === house);
-          studentsOnDom('#students', filter, house);
-        }
-    }
-  });
+const renderToDOM = (divId, content) => {
+  const selectedDiv = document.querySelector(divId);
+  selectedDiv.innerHTML = content;
 };
 
 // ********** HTML Components  ********** //
 // the basic HMTL structure of app
 const htmlStructure = () => {
-    const domString = `
+  const domString = `
     <div id="header-container" class="header mb-3"></div>
     <div id="form-container" class="container mb-3 text-center"></div>
     <div id="filter-container" class="container mb-3"></div>
     <div id="student-container" class="container d-flex"></div>
     `;
 
-  renderToDOM('#app', domString)
+  renderToDOM('#app', domString);
 };
 
 const header = () => {
@@ -85,15 +51,15 @@ const startSortingBtn = () => {
 
 const studentAreas = () => {
   const domString = `<div id="students">No Students</div>
-  <div id="voldy">No Death Eaters</div>`
+  <div id="voldy">No Death Eaters</div>`;
 
   renderToDOM('#student-container', domString);
 };
 
 const studentsOnDom = (divId, array, house = 'Hogwarts') => {
   let domString = '';
-  if(!array.length){
-    domString += `NO ${house.toUpperCase()} STUDENTS`
+  if (!array.length) {
+    domString += `NO ${house.toUpperCase()} STUDENTS`;
   }
 
   array.forEach((student) => {
@@ -153,14 +119,6 @@ const sortStudent = (e) => {
 };
 
 // Create a new ID for the students
-const createId = (array) => {
-  if (array.length) {
-    const idArray = array.map((el) => el.id);
-    return Math.max(...idArray) + 1;
-  }
-  return 0;
-};
-
 
 // add form to DOM on start-sorting click.
 // Add events for form after the form is on the DOM
@@ -184,32 +142,51 @@ const form = () => {
   document.querySelector('#sorting').addEventListener('submit', sortStudent);
 };
 
-const renderToDOM = (divId, content) => {
-  const selectedDiv = document.querySelector(divId);
-  selectedDiv.innerHTML = content;
+const events = () => {
+  // get form on the DOM on button click
+  document.querySelector('#start-sorting').addEventListener('click', () => {
+    // put html elements on the DOM on click
+    form(); // form
+    filterBtnRow(); // filter buttons
+    studentAreas(); // students and voldy's army divs
+  });
+
+  // target expel buttons to move to voldys army
+  document
+    .querySelector('#student-container')
+    .addEventListener('click', (e) => {
+      if (e.target.id.includes('expel')) {
+        const [, id] = e.target.id.split('--');
+        const index = students.findIndex((student) => student.id === Number(id));
+
+        // move from one array to another
+        voldysArmy.push(...students.splice(index, 1));
+        // get both sets of students on the DOM
+        studentsOnDom('#students', students);
+        studentsOnDom('#voldy', voldysArmy);
+      }
+    });
+
+  // target filter buttons on Dom
+  document.querySelector('#filter-container').addEventListener('click', (e) => {
+    if (e.target.id.includes('filter')) {
+      const [, house] = e.target.id.split('--');
+
+      if (house === 'all') {
+        studentsOnDom('#students', students);
+      } else if (house) {
+        const filter = students.filter((student) => student.house === house);
+        studentsOnDom('#students', filter, house);
+      }
+    }
+  });
+};
+
+const startApp = () => {
+  htmlStructure(); // always load first
+  header();
+  startSortingBtn();
+  events(); // always load last
 };
 
 startApp();
-
-const houses = [
-  {
-    house: 'gryffindor',
-    crest:
-      'https://static.wikia.nocookie.net/pottermore/images/1/16/Gryffindor_crest.png'
-  },
-  {
-    house: 'slytherin',
-    crest:
-      'https://static.wikia.nocookie.net/pottermore/images/4/45/Slytherin_Crest.png'
-  },
-  {
-    house: 'hufflepuff',
-    crest:
-      'https://static.wikia.nocookie.net/pottermore/images/5/5e/Hufflepuff_crest.png'
-  },
-  {
-    house: 'ravenclaw',
-    crest:
-      'https://static.wikia.nocookie.net/pottermore/images/4/4f/Ravenclaw_crest.png'
-  }
-];
